+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Breaking: Removes the deprecated APIs. Check the migration guide referenced in our README
 * Adds the ability to opt out of iOS source maps auto upload through the INSTABUG_SOURCEMAPS_UPLOAD_DISABLE env variable
 * Adds dynamic entry file support through the INSTABUG_ENTRY_FILE env variable
+* Adds string keys for the Repro Steps
 * Fixes an issue with setRequestFilterExpression API not working with Hermes
 * Fixes an issue with swipe invocation event not working on Android